refactor: move starknet trees logic to starknet crate
