+++ crates/committer/src/patricia_merkle_tree/filled_tree/errors.rs
@@ -1,9 +1,8 @@
 use tokio::task::JoinError;
 
-use crate::block_committer::input::StarknetStorageValue;
-use crate::patricia_merkle_tree::filled_tree::node::{CompiledClassHash, FilledNode};
+use crate::patricia_merkle_tree::filled_tree::node::FilledNode;
 use crate::patricia_merkle_tree::node_data::errors::LeafError;
-use crate::patricia_merkle_tree::node_data::leaf::{ContractState, Leaf};
+use crate::patricia_merkle_tree::node_data::leaf::Leaf;
 use crate::patricia_merkle_tree::types::NodeIndex;
 use crate::patricia_merkle_tree::updated_skeleton_tree::errors::UpdatedSkeletonTreeError;
 
@@ -28,7 +27,3 @@ pub enum FilledTreeError<L: Leaf> {
     #[error(transparent)]
     JoinError(#[from] JoinError),
 }
-
-pub type StorageTrieError = FilledTreeError<StarknetStorageValue>;
-pub type ClassesTrieError = FilledTreeError<CompiledClassHash>;
-pub type ContractsTrieError = FilledTreeError<ContractState>;
