+++ crates/committer/src/patricia_merkle_tree/node_data.rs
@@ -1,4 +1,3 @@
 pub mod errors;
 pub mod inner_node;
 pub mod leaf;
-pub mod leaf_serde;
