+++ crates/committer/src/patricia_merkle_tree/node_data/leaf.rs
@@ -3,10 +3,7 @@ use std::fmt::Debug;
 use std::future::Future;
 use std::sync::Arc;
 
-use crate::block_committer::input::StarknetStorageValue;
 use crate::felt::Felt;
-use crate::hash::hash_trait::HashOutput;
-use crate::patricia_merkle_tree::filled_tree::node::{ClassHash, CompiledClassHash, Nonce};
 use crate::patricia_merkle_tree::node_data::errors::{LeafError, LeafResult};
 use crate::patricia_merkle_tree::types::NodeIndex;
 use crate::storage::db_object::{DBObject, Deserializable};
@@ -37,54 +34,6 @@ pub trait Leaf: Clone + Sync + Send + DBObject + Deserializable + Default + Debu
             .ok_or(LeafError::MissingLeafModificationData(*index))?
             .clone();
         Ok(leaf_data)
-    }
-}
-
-#[derive(Clone, Debug, Default, Eq, PartialEq)]
-pub struct ContractState {
-    pub nonce: Nonce,
-    pub storage_root_hash: HashOutput,
-    pub class_hash: ClassHash,
-}
-
-impl Leaf for StarknetStorageValue {
-    fn is_empty(&self) -> bool {
-        self.0 == Felt::ZERO
-    }
-
-    async fn create(
-        index: &NodeIndex,
-        leaf_modifications: Arc<LeafModifications<Self>>,
-    ) -> LeafResult<Self> {
-        Self::from_modifications(index, leaf_modifications)
-    }
-}
-
-impl Leaf for CompiledClassHash {
-    fn is_empty(&self) -> bool {
-        self.0 == Felt::ZERO
-    }
-
-    async fn create(
-        index: &NodeIndex,
-        leaf_modifications: Arc<LeafModifications<Self>>,
-    ) -> LeafResult<Self> {
-        Self::from_modifications(index, leaf_modifications)
-    }
-}
-
-impl Leaf for ContractState {
-    fn is_empty(&self) -> bool {
-        self.nonce.0 == Felt::ZERO
-            && self.class_hash.0 == Felt::ZERO
-            && self.storage_root_hash.0 == Felt::ZERO
-    }
-
-    async fn create(
-        index: &NodeIndex,
-        leaf_modifications: Arc<LeafModifications<Self>>,
-    ) -> LeafResult<Self> {
-        Self::from_modifications(index, leaf_modifications)
     }
 }
