+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/config.rs
@@ -1,7 +1,4 @@
-use crate::block_committer::input::StarknetStorageValue;
-use crate::patricia_merkle_tree::filled_tree::node::CompiledClassHash;
-use crate::patricia_merkle_tree::node_data::leaf::{ContractState, Leaf, LeafModifications};
-use crate::patricia_merkle_tree::original_skeleton_tree::errors::OriginalSkeletonTreeError;
+use crate::patricia_merkle_tree::node_data::leaf::Leaf;
 use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTreeResult;
 use crate::patricia_merkle_tree::types::NodeIndex;
 
@@ -11,7 +8,7 @@ pub trait OriginalSkeletonTreeConfig<L: Leaf> {
     /// warning when encountering a trivial modification.
     fn compare_modified_leaves(&self) -> bool;
 
-    /// Compares the previous leaf to the modified and returns true iff they are equal.
+    /// Compares the previous leaf to the modified and returns true if they are equal.
     fn compare_leaf(
         &self,
         index: &NodeIndex,
@@ -23,14 +20,17 @@ pub trait OriginalSkeletonTreeConfig<L: Leaf> {
 macro_rules! generate_trie_config {
     ($struct_name:ident, $leaf_type:ty) => {
         pub struct $struct_name<'a> {
-            modifications: &'a LeafModifications<$leaf_type>,
+            modifications:
+                &'a $crate::patricia_merkle_tree::node_data::leaf::LeafModifications<$leaf_type>,
             compare_modified_leaves: bool,
         }
 
         impl<'a> $struct_name<'a> {
             #[allow(dead_code)]
             pub fn new(
-                modifications: &'a LeafModifications<$leaf_type>,
+                modifications: &'a $crate::patricia_merkle_tree::node_data::leaf::LeafModifications<
+                    $leaf_type,
+                >,
                 compare_modified_leaves: bool,
             ) -> Self {
                 Self { modifications, compare_modified_leaves }
@@ -55,30 +55,4 @@ macro_rules! generate_trie_config {
             }
         }
     };
-}
-
-generate_trie_config!(OriginalSkeletonStorageTrieConfig, StarknetStorageValue);
-
-generate_trie_config!(OriginalSkeletonClassesTrieConfig, CompiledClassHash);
-
-pub(crate) struct OriginalSkeletonContractsTrieConfig;
-
-impl OriginalSkeletonTreeConfig<ContractState> for OriginalSkeletonContractsTrieConfig {
-    fn compare_modified_leaves(&self) -> bool {
-        false
-    }
-
-    fn compare_leaf(
-        &self,
-        _index: &NodeIndex,
-        _previous_leaf: &ContractState,
-    ) -> OriginalSkeletonTreeResult<bool> {
-        Ok(false)
-    }
-}
-
-impl OriginalSkeletonContractsTrieConfig {
-    pub(crate) fn new() -> Self {
-        Self
-    }
 }
