+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/node.rs
@@ -3,7 +3,7 @@ use crate::patricia_merkle_tree::node_data::inner_node::PathToBottom;
 
 #[derive(Copy, Clone, Debug, PartialEq, Eq)]
 /// A node in the structure of a Patricia-Merkle tree, before the update.
-pub(crate) enum OriginalSkeletonNode {
+pub enum OriginalSkeletonNode {
     Binary,
     Edge(PathToBottom),
     // Represents a root of a subtree where none of it's descendants has changed.
