+++ crates/committer/src/patricia_merkle_tree/updated_skeleton_tree/node.rs
@@ -3,7 +3,7 @@ use crate::patricia_merkle_tree::node_data::inner_node::PathToBottom;
 
 /// A node in the structure of a Patricia-Merkle tree, after the update.
 #[derive(Debug, Clone, PartialEq)]
-pub(crate) enum UpdatedSkeletonNode {
+pub enum UpdatedSkeletonNode {
     Binary,
     Edge(PathToBottom),
     // Represents a root of a subtree where none of it's descendants has changed.
