+++ crates/starknet-committer/src/lib.rs
@@ -1,2 +1,2 @@
 pub mod block_committer;
-
+pub mod starknet_patricia_merkle_tree;
