+++ crates/starknet-committer/src/starknet_patricia_merkle_tree.rs
@@ -1,0 +1,4 @@
+pub mod node;
+pub mod starknet_leaf;
+pub mod tree;
+pub mod types;
