+++ crates/starknet-committer/src/starknet_patricia_merkle_tree/starknet_leaf.rs
@@ -1,0 +1,4 @@
+pub mod leaf;
+pub mod leaf_serde;
+#[cfg(test)]
+pub mod leaf_serde_test;
